+++ src/browser/settings.jsx
@@ -61,6 +61,7 @@ var SettingsPage = React.createClass({
       trayIconTheme: this.state.trayIconTheme,
       disablewebsecurity: this.state.disablewebsecurity,
       version: settings.version,
+      minimizeToTray: this.state.minimizeToTray,
       toggleWindowOnTrayIconClick: this.state.toggleWindowOnTrayIconClick,
       notifications: {
         flashWindow: this.state.notifications.flashWindow
@@ -101,9 +102,16 @@ var SettingsPage = React.createClass({
     });
   },
   handleChangeShowTrayIcon: function() {
+    var shouldShowTrayIcon = this.refs.showTrayIcon.getChecked();
     this.setState({
-      showTrayIcon: this.refs.showTrayIcon.getChecked()
+      showTrayIcon: shouldShowTrayIcon
     });
+
+    if (process.platform === 'darwin' && !shouldShowTrayIcon) {
+      this.setState({
+        minimizeToTray: false
+      });
+    }
   },
   handleChangeTrayIconTheme: function() {
     this.setState({
@@ -113,6 +121,14 @@ var SettingsPage = React.createClass({
   handleChangeAutoStart: function() {
     this.setState({
       autostart: this.refs.autostart.getChecked()
+    });
+  },
+  handleChangeMinimizeToTray: function() {
+    var shouldMinimizeToTray = (process.platform !== 'darwin' || this.refs.showTrayIcon.getChecked())
+    && this.refs.minimizeToTray.getChecked();
+
+    this.setState({
+      minimizeToTray: shouldMinimizeToTray
     });
   },
   handleChangeToggleWindowOnTrayIconClick: function() {
@@ -162,6 +178,11 @@ var SettingsPage = React.createClass({
     if (process.platform === 'win32' || process.platform === 'linux') {
       options.push(<Input key="inputAutoStart" id="inputAutoStart" ref="autostart" type="checkbox" label="Start app on login." checked={ this.state.autostart } onChange={ this.handleChangeAutoStart }
                    />);
+    }
+
+    if (process.platform === 'darwin') {
+      options.push(<Input key="inputMinimizeToTray" id="inputMinimizeToTray" ref="minimizeToTray" type="checkbox" label="Leave app running in notification area when the window is closed"
+                     disabled={ !this.state.showTrayIcon } checked={ this.state.minimizeToTray } onChange={ this.handleChangeMinimizeToTray } />);
     }
 
     if (process.platform === 'win32') {