+++ test/specs/browser/settings_test.js
@@ -151,6 +151,16 @@ describe('browser/settings.html', function() {
       });
     });
 
+    describe('Minimize to tray', function() {
+      it('should appear on darwin', function() {
+        const expected = (process.platform === 'darwin');
+        env.addClientCommands(this.app.client);
+        return this.app.client
+          .loadSettingsPage()
+          .isExisting('#inputMinimizeToTray').should.eventually.equal(expected)
+      });
+    });
+
     describe('Toggle window visibility when clicking on the tray icon', function() {
       it('should appear on win32', function() {
         const expected = (process.platform === 'win32');